DO NOT MERGE.  Integrate support work from master.
First submit of FragmentPager class.

This provides an easy way to build a UI where the user can
swipe left or right through its elements.  The elements are
implemented as fragments, and the class takes care of managing
those fragments as the user navigates through it.

This implementation also introduces a new FragmentManager
concept of a "detached" fragment -- basically a way for you
to put a fragment in the same state as when it is on the
back stack, where the framework is managing its current state
but it is no longer actively running.

Also required the introduction of new compatibility code for
MotionEvent and VelocityTracker for accessing multi-touch data.

Tweak view save/restore state so it will play well with list views.

We need to restore the state *after* the adapter has been set;
setting the adapter clears the state.  To do this, we move the
state restore from immediately after the view is created to after
we call Fragment.onActivityAttached().

Also introduced a new Fragment.onViewCreated() callback which is
nice for fragments that want to do some setup after onCreateView()
but allow for subclasses to override that method.  (ListFragment
I'm looking at you.)