First submit of FragmentPager class.
This provides an easy way to build a UI where the user can
swipe left or right through its elements.  The elements are
implemented as fragments, and the class takes care of managing
those fragments as the user navigates through it.

This implementation also introduces a new FragmentManager
concept of a "detached" fragment -- basically a way for you
to put a fragment in the same state as when it is on the
back stack, where the framework is managing its current state
but it is no longer actively running.

Also required the introduction of new compatibility code for
MotionEvent and VelocityTracker for accessing multi-touch data.

Change-Id: I997492cfc6f1969f1a5d775385afb830f5a47795