Fix issue #5085211: Leaked cursors in Loaders
It looks like canceling of async tasks on froyo behaves differently
than newer platforms.  To deal with this, we'll just have our own
impl of the async task engine in the support library for use on those
older platforms.

It would be nice to switch to the built-in AsyncTask when running on
newer versions of the platform...  not sure the best way to do that,
though.

Change-Id: Ic2209c83033b76f12be5ceee0809f4fc57af7026