Bug 5300610 - Swiping between pages of a ViewPager can be slow
Add setOffscreenPageLimit to ViewPager. This allows apps to specify
how many pages will be kept active to either side of the current page.

Normally this is not needed, but the pathological case for ViewPager
is an adapter with a fixed set of 3 pages when this limit is set to 1.
Every swipe from an edge will page in the furthest item, forcing it to
relayout. setOffscreenPageLimit allows these apps to keep more pages
active and skip extra layout passes.

Change-Id: I319181ce354b9004f1b7a26cbfd0c1c399cb293f