Fix NPE in `onActivityResult` when target fragment no longer exists.
If an activity result comes in which originated from a fragment that
can no longer be found an attempt is still made to call a method
on the instance. The check for null already existed to produce a
logging statement but the subsequent method call was not placed
into an `else` result in the attempt to call being made every time.

Change-Id: I0378d516a16e511a6195608243c049b1cb50a8c8