Add GridLayout as a support project library.
The library is distributed as a project containing a single jar and resources
but the jar file is built by the Android build system in order to target the
current API.
The code also relies on the R class being generated, which is currently done
manually. The build system uses an intermediary library to compile the R class
so that it doesn't get packaged in the final jar file (same behavior as
compiling the library project with the SDK).

Change-Id: I36ce97d875f2efbb50cca665e9fd0e668f5ba837