Add paging gutters to ViewPager.
With the new Android Design guidelines, conforming apps pair tabs with
a ViewPager to easily move between tabbed content. However taken to
its natural conclusion this limits the types of content that can be
placed into a tab. Unbounded scrolling content such as a map suddenly
cannot be used, since there is no way to reach the "edge" of that
content in order to start paging. In addition, bounded but very wide
horizontally scrolling content can make paging cumbersome.

Enter paging gutters.

A gutter is defined as the 16dp space along the left and right edges
of a ViewPager, or 10% of the available horizontal space, whichever is
smaller. If the user begins a drag within a gutter, the pager will
take over and allow the user to switch pages without reaching the edge
of horizontally scrolling content. This provides a default "edge
swipe" navigation that may be used in any ViewPager regardless of
content.

Bug 6486971

Change-Id: I27a393445e7d379204b42fd3751acd4c1b4fa69a