Add PageTransformers to ViewPager
PageTransformer allows an app to supply a custom step during scrolling
to transform each page view in addition to the default sliding
behavior. Setting a PageTransformer also allows the app to reverse the
drawing order for pages to accomodate certain kinds of
transformations.

Add setLayerType to ViewCompat. Make ViewPager use hardware layers for
scrolling when available and a PageTransformer is in use.

Change-Id: Icdcb7bf92ad4adb9d14e36c6048f7f5dd455437b