Make ActionBarImplCompat depend on FragmentActivity
It previously depended on ActionBarActivity. However, the generic
"java" package is built after the "eclair" pacakge, so this class
isn't available.

Swapping the build around won't work, since we'll need the "java"
package to come last for class/method dispatch. The good news is,
it seems this can be done easily without breaking anything.

Change-Id: I67b70af778fa3a85a714fb6ae759a74bbf994d55