Fix how fragment/tab interaction works to avoid list view state loss.
The implementation of FragmentTabHost would cause us to toggle between
the default and selected tab fragments when restoring our state.
Because of an issue in ListView (fixed by another change), this would
cause us to lose the state because ListView tripped over itself if
you asked it to save its state immediately after restoring it.  We
are now more careful about switching fragments, so we don't run in to
this problem at least in most cases.

Also pull over framework change to not let inactive fragments be
more than CREATED.

Bug #7232088: ListView saved state being lost in some cases

Change-Id: Id9f20c207409d2479ff8bce8ed5b0553510cdd63