Move support Menu interfaces + callbacks to v4 (DO NOT MERGE)
This moves the Menu* interfaces used for the ActionBar
into android-support-v4. This allows us to remove
ActionBarFragment and related hackery.

This commit just keeps us at the same level of functionality,
but is a cleaner implementation. A future feature would be
to enable the use of the support Menu when not using
ActionBarActivity.

Change-Id: I952ce78a378e8ec3b990f5c285a44c7fc4807f0c
Signed-off-by: Chris Banes <chrisbanes@google.com>