DrawerLayout features/bugfixes/tuning/optimization
Add the ability to draw shadows on the edges of drawers.

Further optimize drawing (or not drawing) of drawers. Clip the content
view if an area is covered by an opaque drawer.

Mark offscreen drawer views as invisible.

Fix a bug where an edge touch would peek (and thereby close) a drawer.

Change-Id: I6ae12ee9d7118aa3209aeb93ead86ca5c9292f82