Add media router and display manager to support library.
The support library MediaRouter API is somewhat different from
the framework's existing MediaRouter API.  It is designed to
be simpler and easier to extend.

In the new MediaRouter, routes are published by MediaRouteProviders.
MediaRouteProviders replace the concept of UserRoutes.
The built-in SystemMediaRouteProvider publishes the default
system routes and interoperates with the framework MediaRouter.

Each route now declares its capabilities and behaviors by way
of MediaControlIntents.  The type of the route (live audio,
live video, or remote playback) is specified by the categories
of media control intents that the route supports.  A route
may support any number of custom provider-defined media control
intent categories that provide access to specialized features.

The old concepts of route categories and route groups have
been removed.  Route providers are now responsible for implementing
the route grouping and configuration through some UI of their
own that is not managed by the media router.

This patch does not include support for route icon drawables,
the new route picker UI or route provider services.  These features
will be added in follow on patches.

Change-Id: I10cf078917d053a916254a5aa9d208b8cfd341ea