Tidy up Delegates + Enforce AppCompat theme usage
This commit moves all variables and state into the relevant
delegate classes. This means that ActionBarActivity is now
merely a call proxy and entry point.

As we now have a common class amongst the delegates, I also
fixed the issue where we do not enforce the usage of AppCompat
styles. This will fix issue 8124469.

Change-Id: If055906e754ceda471260f3160009cb384703ecd
Signed-off-by: Chris Banes <chrisbanes@google.com>