Fix Home Action Bar Item
At the moment the Home action bar doesn't actually work,
the necessary code was commented out as we can't use the
same method as the framework.

So I've created a Callback interface in ActionBarView,
which it calls when the home icon is clicked.

At the same time I've changed the home menu id to be
android.R.id.home so devs don't have to catch two
different ids.

Change-Id: If3b67e437b993466ffd158392680fcea0e4ca79f
Signed-off-by: Chris Banes <chrisbanes@google.com>