Rename Resources to have abc_ prefix
This commit renames all exposed resources to
have the abc_ prefix.

This is so that ActionBarCompat's resources
do not interfere with any Holo resources that
devs may import themsevles. This is also
if devs use tools such as 'Action Bar Style
Generator'.

This allow devs to quick see which resources
are from ABC.

Change-Id: Ib9597110469c94ee3b70c9e89d0640033aa477f9
Signed-off-by: Chris Banes <chrisbanes@google.com>