Fix List Menu text styling on pre-HC devices
This commit fixes the styling on text in list menu items.

To combat the different panel background which can be present
on devices, we set our own (imported from Holo). The panel
background changes with the theme, similar to Holo styles. This
means we can assume a particular background, and set an appropriate
text colour. Even with the drawables our drawable size is ~43KB.

This commit also changes the text colour on various widgets to use the
disable-only colour. This fixes a bug on dark styles where the text
colour would change if selected/clicked.

Change-Id: I90deb13bc5fd435710bca6bf1e72a7af0c217a89
Signed-off-by: Chris Banes <chrisbanes@google.com>