ViewPager optimization/fix for pathological case
Prune/don't generate pages to fill the desired full width to either
side if the measured width of the pager is 0. This fixes a bug in some
apps that would measure their ViewPager at width 0 in an initial pass
when placed in some layouts, resulting in too many instantiateItem
calls.

Change-Id: I87d036814b53756cac49d0a68ca9e45115f6c3b2