Fix Theme.AppCompat.Light.DarkActionBar styling.
The base themes need to propagate appcompat attributes such as
actionBarWidgetTheme back into their framework counterparts to
ensure that the correct styles are applied.

This change specifically fixes one such problem but there will be
others to be resolved later as part of this bug.

Fix visibility of a couple of overridden methods in
ActionBarActivity to match that of the base class.

Bug: 9196417
Change-Id: I0799f22234109b1d6f4ef8c1bb39bc7530054c06