Refine SlidingPaneLayout opacity detection
Before API 18, View#isOpaque would aggressively return false for
scrollbar modes that still implied view opacity and SlidingPaneLayout
relied directly on the format of the background drawable as a
fallback. In API 18 and above, this has been fixed.

Use View#isOpaque as the primary signal and only fall back on the
background drawable check on older platform versions if isOpaque
returns false. This allows applications with custom views to override
isOpaque as needed when they're Really Really Sure that they want
SlidingPaneLayout's detected opacity behavior instead of creating a
Drawable that lies.

Bug 9466551

Change-Id: I668f0a79fe83e09a36878d4c2f8b77258d55af6b