Adding a code name version check until the SDK constant for K is set.
The code in PrintHelper was using only the SDK version to determine
whether to use the K specific print code. While this is correct we
still do not have the SDK constant for K set and the incorrect
implementation is loaded. This change adds a check for the code name
or the SDK version to ensure it works now and in the future. There
is a bug tracking the removal of the code name check.

bug:11099831

Change-Id: Ib0b6ed355519385957a7a3a53de507ba7d866687