Add typedef enum and resource type annotations
The @IntDef and @StringDef annotations can list the specific
enum or flag constants that can be supplied for a given int or
String parameter or return value.

The Res annotations can be used to indicate that a given int
parameter or return value must correspond to a resource of the
given type.

Change-Id: I8d740360d3edcef44a6d1a053cf16da53f34a0ee